Abstract out plane chunk mesh generation
